Moves `UfmPlugin` type to its own referencable file
