+++ src/Umbraco.Web.UI.Client/src/packages/ufm/plugins/index.ts
@@ -1,2 +1,1 @@
 export { ufm } from './marked-ufm.plugin.js';
-export type { UfmPlugin, UfmToken } from './marked-ufm.plugin.js';
