+++ src/Umbraco.Web.UI.Client/src/packages/ufm/plugins/marked-ufm.plugin.ts
@@ -1,18 +1,7 @@
-import type { MarkedExtension, Tokens } from '@umbraco-cms/backoffice/external/marked';
-
-export interface UfmPlugin {
-	alias: string;
-	marker?: string;
-	render?: (token: UfmToken) => string | undefined;
-}
-
-export interface UfmToken extends Tokens.Generic {
-	prefix: string;
-	text?: string;
-}
+import type { UfmPlugin } from './types.js';
+import type { MarkedExtension } from '@umbraco-cms/backoffice/external/marked';
 
 /**
- *
  * @param {Array<UfmPlugin>} plugins - An array of UFM plugins.
  * @returns {MarkedExtension} A Marked extension object.
  */
