+++ src/Umbraco.Web.UI.Client/src/packages/ufm/plugins/types.ts
@@ -1,0 +1,12 @@
+import type { Tokens } from '@umbraco-cms/backoffice/external/marked';
+
+export interface UfmPlugin {
+	alias: string;
+	marker?: string;
+	render?: (token: UfmToken) => string | undefined;
+}
+
+export interface UfmToken extends Tokens.Generic {
+	prefix: string;
+	text?: string;
+}
