+++ src/Umbraco.Web.UI.Client/src/packages/ufm/types.ts
@@ -1,1 +1,2 @@
 export type * from './extensions/types.js';
+export type * from './plugins/types.js';
